Update paths inside our component stories after moving the static folder
