+++ stories/avatars.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { boolean, number, select } from '@storybook/addon-knobs';
 import { Avatar, AvatarStack, Bullet, Counter, TextBody, Tooltip } from '../components';
-import avatars from '../static/data/avatar';
+import avatars from './static/data/avatar';
 
 const directions = ['horizontal', 'vertical'];
 const sizes = ['tiny', 'small', 'medium'];
