+++ stories/tab.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { boolean } from '@storybook/addon-knobs/react';
 import { IconTab, TabGroup, TitleTab, Box, Counter as UICounter } from '../components';
-import { tabItems, invertedTabItems } from '../static/data/tab';
+import { tabItems, invertedTabItems } from './static/data/tab';
 
 import {
   IconCalendarMediumOutline,
